Extract styled-component definitions from primary module files
