+++ src/components/App/demo.js
@@ -2,6 +2,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import App from './App';
+import App from '.';
 
 storiesOf('App', module).add('default', () => <App>App goes here</App>);
