+++ src/components/App/index.js
@@ -1,0 +1,4 @@
+// @flow
+import App from './App';
+
+export default App;
