+++ src/components/App/test.js
@@ -2,9 +2,9 @@
 /* eslint-env jest */
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Main from './Main';
+import App from '.';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<Main />, div);
+  ReactDOM.render(<App />, div);
 });
