+++ src/components/Footer/demo.js
@@ -2,6 +2,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import Footer from './Footer';
+import Footer from '.';
 
 storiesOf('Footer', module).add('default', () => <Footer />);
