+++ src/components/Footer/test.js
@@ -2,9 +2,9 @@
 /* eslint-env jest */
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import Footer from '.';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<App />, div);
+  ReactDOM.render(<Footer />, div);
 });
