+++ src/components/Header/Root.js
@@ -2,5 +2,5 @@
 import styled from 'styled-components';
 
 export default styled.div`
-  background-color: sandybrown;
+  background-color: lightblue;
 `;
