+++ src/components/Header/demo.js
@@ -2,6 +2,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import Header from './Header';
+import Header from '.';
 
 storiesOf('Header', module).add('default', () => <Header />);
