+++ src/components/Header/index.js
@@ -1,8 +1,7 @@
 // @flow
 import React from 'react';
-import styled from 'styled-components';
 
-const Root = styled.div`background-color: lightblue;`;
+import Root from './Root';
 
 const Header = () => <Root>Header goes here</Root>;
