+++ src/components/Header/test.js
@@ -2,9 +2,9 @@
 /* eslint-env jest */
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Footer from './Footer';
+import Header from '.';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<Footer />, div);
+  ReactDOM.render(<Header />, div);
 });
