+++ src/components/Main/Wrapper.js
@@ -2,5 +2,5 @@
 import styled from 'styled-components';
 
 export default styled.div`
-  background-color: sandybrown;
+  background-color: beige;
 `;
