+++ src/components/Main/test.js
@@ -2,9 +2,9 @@
 /* eslint-env jest */
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Header from './Header';
+import Main from '.';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<Header />, div);
+  ReactDOM.render(<Main />, div);
 });
