+++ src/components/Sidebar/demo.js
@@ -2,6 +2,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import Sidebar from './Sidebar';
+import Sidebar from '.';
 
 storiesOf('Sidebar', module).add('default', () => <Sidebar />);
