+++ src/components/Sidebar/test.js
@@ -2,7 +2,7 @@
 /* eslint-env jest */
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Sidebar from './Sidebar';
+import Sidebar from '.';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
