+++ src/components/index.js
@@ -1,6 +1,6 @@
 // @flow
-export { default as App } from './App/App';
-export { default as Footer } from './Footer/Footer';
-export { default as Main } from './Main/Main';
-export { default as Header } from './Header/Header';
-export { default as Sidebar } from './Sidebar/Sidebar';
+export { default as App } from './App';
+export { default as Footer } from './Footer';
+export { default as Main } from './Main';
+export { default as Header } from './Header';
+export { default as Sidebar } from './Sidebar';
